pyproject.toml
[project]
name = "nocode"
version = "0.2.0"
description = "NoCode AI agent companion"
requires-python = ">=3.12"
dependencies = [
    "agent-client-protocol>=0.8.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "mcp>=1.27.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.42.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
nocode = "nocode_agent.main:main"
nocode-acp = "nocode_agent.acp_server:main"
nocode-web = "multiagent_system.server:main"

[tool.setuptools]
packages = ["nocode_agent", "multiagent_system"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
